Added a Nash-Cournot simulator
This expands a bit on the Keen and Standish 2010 model (which claims
that firms tend to monopoly) by severing the optimization procedure to
only allowing a single firm to optimize each period.  This tends to the
Cournot, not monopoly, result.

This suggests that K&S's model is actually specifying a model with a
form of tacit collusion: since firms are following a strategy of basing
each action on the previous period's action, there will tend to be an
increased correlation between actions (since correlated actions are
more likely to increase profits).  Thus the positive or negative aspect
of quantity adjustments becomes increasingly correlated over time, to
the point that increasingly, firms are all taking positive (or
negative) steps at the same time: and with such a correlation, it's
clear why the results tend towards monopoly.